+++ src/LearnNavigation/Screens.js
@@ -1,0 +1,5 @@
+// @flow
+
+export const WELCOME_SCREEN = 'Learn.Welcome';
+export const LOGIN_SCREEN = 'Learn.Login';
+export const DASHBOARD_SCREEN = 'Learn.Dashboard';
