+++ src/LearnNavigation/index.js
@@ -1,0 +1,8 @@
+export * from './Screens';
+export { pushAuthenticated } from './Authenticated';
+export { pushUnauthenticated } from './Unauthenticated'
+
+import registerScreens from './registerScreens';
+
+// Register all screens on launch
+registerScreens();
