+++ src/LearnNavigation/registerScreens.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import { Navigation } from 'react-native-navigation';
 
-import { WelcomeScreen, LoginScreen, DashboardScreen } from 'AppScreens';
+import { WelcomeScreen, LoginScreen, DashboardScreen } from 'LearnScreens';
 import { Provider } from 'AppRedux';
 
 import { WELCOME_SCREEN, LOGIN_SCREEN, DASHBOARD_SCREEN } from './Screens';
@@ -26,5 +26,4 @@ export default function () {
   Navigation.registerComponent(WELCOME_SCREEN, () => WrappedComponent(WelcomeScreen));
   Navigation.registerComponent(LOGIN_SCREEN, () => WrappedComponent(LoginScreen));
   Navigation.registerComponent(DASHBOARD_SCREEN, () => WrappedComponent(DashboardScreen));
-  console.info('All screens have been registered...');
 }
