+++ src/LearnScreens/WelcomeScreen/__tests__/WelcomeScreen.test.js
