--- src/navigation/Screens.js
@@ -1,5 +1,0 @@
-// @flow
-
-export const WELCOME_SCREEN = 'demo.WelcomeScreen';
-export const LOGIN_SCREEN = 'demo.LoginScreen';
-export const DASHBOARD_SCREEN = 'demo.DashboardScreen';
