--- src/navigation/index.js
@@ -1,2 +1,0 @@
-export * from './Screens';
-export { pushTutorialScreen, pushDashboard } from './Navigation';
