--- src/navigation/package.json
@@ -1,3 +1,0 @@
-{
-  "name": "AppNavigator"
-}
